pyproject.toml
[project]
name = "ai-interviewer-livekit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-sarvam~=1.0",
    "python-dotenv",
    "numpy<2.0",
    "gradio>=3.36.1",
    "flask>=3.1.2",
    "pyjwt>=2.10.1",
    "livekit-plugins-deepgram>=1.3.10",
    "livekit-plugins-cartesia>=1.3.10",
    "livekit-plugins-turn-detector>=1.3.10",
    "pypdf>=6.6.0",
    "livekit-api>=1.1.0",
]
